+++ marshmallow/marshalling.py
@@ -14,8 +14,6 @@ from marshmallow import base, utils
 from marshmallow.compat import text_type, iteritems
 from marshmallow.exceptions import (
     ValidationError,
-    MarshallingError,
-    UnmarshallingError,
 )
 
 __all__ = [
@@ -42,10 +40,7 @@ class _Missing(_Null):
         return '<marshmallow.marshalling.missing>'
 
 
-# Singleton that represents an empty value. Used as the default for Nested
-# fields so that `Field._call_with_validation` is invoked, even when the
-# object to serialize has the nested attribute set to None. Therefore,
-# `RegistryErrors` are properly raised.
+# Singleton that represents an empty value.
 null = _Null()
 
 # Singleton value that indicates that a field's value is missing from input
@@ -54,63 +49,64 @@ null = _Null()
 missing = _Missing()
 
 
-def _call_and_store(getter_func, data, field_name, field_obj, errors_dict,
-               exception_class, strict=False, index=None):
-    """Helper method for DRYing up logic in the :meth:`Marshaller.serialize` and
-    :meth:`Unmarshaller.deserialize` methods. Call ``getter_func`` with ``data`` as its
-    argument, and store any errors of type ``exception_class`` in ``error_dict``.
+class ErrorStore(object):
 
-    :param callable getter_func: Function for getting the serialized/deserialized
-        value from ``data``.
-    :param data: The data passed to ``getter_func``.
-    :param str field_name: Field name.
-    :param FieldABC field_obj: Field object that performs the
-        serialization/deserialization behavior.
-    :param dict errors_dict: Dictionary to store errors on.
-    :param type exception_class: Exception class that will be caught during
-        serialization/deserialization. Errors of this type will be stored
-        in ``errors_dict``.
-    :param int index: Index of the item being validated, if validating a collection,
-        otherwise `None`.
-    """
-    try:
-        value = getter_func(data)
-    except exception_class as err:  # Store errors
-        if strict:
-            err.field = field_obj
-            err.field_name = field_name
-            raise err
-        # Warning: Mutation!
-        if index is not None:
-            errors = {}
-            errors_dict[index] = errors
-        else:
-            errors = errors_dict
-        # Warning: Mutation!
-        if (hasattr(err, 'underlying_exception') and
-                isinstance(err.underlying_exception, ValidationError)):
-            validation_error = err.underlying_exception
-            if isinstance(validation_error.messages, dict):
-                errors[field_name] = validation_error.messages
+    def __init__(self):
+        #: Dictionary of errors stored during serialization
+        self.errors = {}
+        #: List of `Field` objects which have validation errors
+        self.error_fields = []
+        #: List of field_names which have validation errors
+        self.error_field_names = []
+        #: True while (de)serializing a collection
+        self._pending = False
+
+    def reset_errors(self):
+        self.errors = {}
+        self.error_field_names = []
+        self.error_fields = []
+
+    def call_and_store(self, getter_func, data, field_name, field_obj, index=None):
+        """Call ``getter_func`` with ``data`` as its argument, and store any `ValidationErrors`.
+
+        :param callable getter_func: Function for getting the serialized/deserialized
+            value from ``data``.
+        :param data: The data passed to ``getter_func``.
+        :param str field_name: Field name.
+        :param FieldABC field_obj: Field object that performs the
+            serialization/deserialization behavior.
+        :param int index: Index of the item being validated, if validating a collection,
+            otherwise `None`.
+        """
+        try:
+            value = getter_func(data)
+        except ValidationError as err:  # Store validation errors
+            self.error_fields.append(field_obj)
+            self.error_field_names.append(field_name)
+            if index is not None:
+                errors = {}
+                self.errors[index] = errors
             else:
-                errors.setdefault(field_name, []).extend(validation_error.messages)
-        else:
-            errors.setdefault(field_name, []).append(text_type(err))
-        value = None
-    except TypeError:
-        # field declared as a class, not an instance
-        if (isinstance(field_obj, type) and
-                issubclass(field_obj, base.FieldABC)):
-            msg = ('Field for "{0}" must be declared as a '
-                            'Field instance, not a class. '
-                            'Did you mean "fields.{1}()"?'
-                            .format(field_name, field_obj.__name__))
-            raise TypeError(msg)
-        raise
-    return value
+                errors = self.errors
+            # Warning: Mutation!
+            if isinstance(err.messages, dict):
+                errors[field_name] = err.messages
+            else:
+                errors.setdefault(field_name, []).extend(err.messages)
+            value = None
+        except TypeError:
+            # field declared as a class, not an instance
+            if (isinstance(field_obj, type) and
+                    issubclass(field_obj, base.FieldABC)):
+                msg = ('Field for "{0}" must be declared as a '
+                                'Field instance, not a class. '
+                                'Did you mean "fields.{1}()"?'
+                                .format(field_name, field_obj.__name__))
+                raise TypeError(msg)
+            raise
+        return value
 
-
-class Marshaller(object):
+class Marshaller(ErrorStore):
     """Callable class responsible for serializing data and storing errors.
 
     :param str prefix: Optional prefix that will be prepended to all the
@@ -118,10 +114,7 @@ class Marshaller(object):
     """
     def __init__(self, prefix=''):
         self.prefix = prefix
-        #: Dictionary of errors stored during serialization
-        self.errors = {}
-        #: True while serializing a collection
-        self.__pending = False
+        ErrorStore.__init__(self)
 
     def serialize(self, obj, fields_dict, many=False, strict=False, skip_missing=False,
                   accessor=None, dict_class=dict, index_errors=True, index=None):
@@ -147,29 +140,26 @@ class Marshaller(object):
             Renamed from ``marshal``.
         """
         # Reset errors dict if not serializing a collection
-        if not self.__pending:
-            self.errors = {}
+        if not self._pending:
+            self.reset_errors()
         if many and obj is not None:
-            self.__pending = True
+            self._pending = True
             ret = [self.serialize(d, fields_dict, many=False, strict=strict,
                                     dict_class=dict_class, accessor=accessor,
                                     skip_missing=skip_missing,
                                     index=idx, index_errors=index_errors)
                     for idx, d in enumerate(obj)]
-            self.__pending = False
+            self._pending = False
             return ret
         items = []
         for attr_name, field_obj in iteritems(fields_dict):
             key = ''.join([self.prefix, attr_name])
             getter = lambda d: field_obj.serialize(attr_name, d, accessor=accessor)
-            value = _call_and_store(
+            value = self.call_and_store(
                 getter_func=getter,
                 data=obj,
                 field_name=key,
                 field_obj=field_obj,
-                errors_dict=self.errors,
-                exception_class=MarshallingError,
-                strict=strict,
                 index=(index if index_errors else None)
             )
             skip_conds = (
@@ -180,22 +170,23 @@ class Marshaller(object):
             if any(skip_conds):
                 continue
             items.append((key, value))
+        if self.errors and strict:
+            raise ValidationError(
+                self.errors,
+                field_names=self.error_field_names,
+                fields=self.error_fields
+            )
         return dict_class(items)
 
     # Make an instance callable
     __call__ = serialize
 
 
-class Unmarshaller(object):
+class Unmarshaller(ErrorStore):
     """Callable class responsible for deserializing data and storing errors.
 
     .. versionadded:: 1.0.0
     """
-    def __init__(self):
-        #: Dictionary of errors stored during deserialization
-        self.errors = {}
-        #: True while deserializing a collection
-        self.__pending = False
 
     def _validate(self, validators, output, raw_data, fields_dict, strict=False):
         """Perform schema-level validation. Stores errors if ``strict`` is `False`.
@@ -214,15 +205,12 @@ class Unmarshaller(object):
                     ))
             except ValidationError as err:
                 # Store or reraise errors
-                if err.fields:
-                    field_names = err.fields
+                if err.field_names:
+                    field_names = err.field_names
                     field_objs = [fields_dict[each] for each in field_names]
                 else:
                     field_names = ['_schema']
                     field_objs = []
-                if strict:
-                    raise UnmarshallingError(err, fields=field_objs,
-                                             field_names=field_names)
                 for field_name in field_names:
                     if isinstance(err.messages, (list, tuple)):
                         # self.errors[field_name] may be a dict if schemas are nested
@@ -236,6 +224,12 @@ class Unmarshaller(object):
                         self.errors.setdefault(field_name, []).append(err.messages)
                     else:
                         self.errors.setdefault(field_name, []).append(text_type(err))
+                if strict:
+                    raise ValidationError(
+                        self.errors,
+                        fields=field_objs,
+                        field_names=field_names
+                    )
         return output
 
     def deserialize(self, data, fields_dict, many=False, validators=None,
@@ -261,16 +255,16 @@ class Unmarshaller(object):
         :return: A dictionary of the deserialized data.
         """
         # Reset errors if not deserializing a collection
-        if not self.__pending:
-            self.errors = {}
+        if not self._pending:
+            self.reset_errors()
         if many and data is not None:
-            self.__pending = True
+            self._pending = True
             ret = [self.deserialize(d, fields_dict, many=False,
                         validators=validators, preprocess=preprocess,
                         postprocess=postprocess, strict=strict, dict_class=dict_class,
                         index=idx, index_errors=index_errors)
                     for idx, d in enumerate(data)]
-            self.__pending = False
+            self._pending = False
             return ret
         raw_data = data
         if data is not None:
@@ -287,14 +281,11 @@ class Unmarshaller(object):
                     raw_value = _miss() if callable(_miss) else _miss
                 if raw_value is missing and not field_obj.required:
                     continue
-                value = _call_and_store(
+                value = self.call_and_store(
                     getter_func=field_obj.deserialize,
                     data=raw_value,
                     field_name=key,
                     field_obj=field_obj,
-                    errors_dict=self.errors,
-                    exception_class=UnmarshallingError,
-                    strict=strict,
                     index=(index if index_errors else None)
                 )
                 if raw_value is not missing:
@@ -311,6 +302,12 @@ class Unmarshaller(object):
             validators = validators or []
             ret = self._validate(validators, ret, raw_data, fields_dict=fields_dict,
                                  strict=strict)
+        if self.errors and strict:
+            raise ValidationError(
+                self.errors,
+                field_names=self.error_field_names,
+                fields=self.error_fields
+            )
         if postprocess:
             postprocess = postprocess or []
             for func in postprocess: